--- packages/components-native/src/CollapsableItem/index.ts
@@ -1,1 +1,0 @@
-export { CollapsableItem } from "./CollapsableItem";
