+++ packages/components-native/src/Disclosure/index.ts
@@ -1,0 +1,1 @@
+export { Disclosure } from "./Disclosure";
