+++ packages/components-native/src/index.ts
@@ -8,8 +8,8 @@ export * from "./Button";
 export * from "./ButtonGroup";
 export * from "./Card";
 export * from "./Chip";
-export * from "./CollapsableItem";
 export * from "./Content";
+export * from "./Disclosure";
 export * from "./Divider";
 export * from "./EmptyState";
 export * from "./ErrorMessageWrapper";
